feat: 로그인 관련 에러 핸들링(#269)

Closes #269

See merge request s11-webmobile1-sub2/S11P12C108!308

+++ meetz-front/src/hooks/meeting/useCheckAuth.tsx
@@ -1,6 +1,6 @@
 import { useNavigate } from "react-router-dom";
 import getAccessTKByRefreshTK from "../../apis/auth/getAccessTKByRefreshTK";
-import { useEffect } from "react";
+import { useEffect, useState } from "react";
 import fetchUserData from "../../lib/fetchUserData";
 import clearUserData from "../../lib/clearUserData";
 import setUserData from "../../lib/setUserData";
@@ -10,36 +10,29 @@ import setUserData from "../../lib/setUserData";
  */
 const useCheckAuth = (pageRole: string) => {
   const navigate = useNavigate();
+  const [isAuthChecked, setIsAuthChecked] = useState(false);
 
   const refreshAccessToken = async (refreshTKOG: string) => {
     try {
       const { accessToken, expireAt, role, refreshToken } = await getAccessTKByRefreshTK(refreshTKOG);
       setUserData(role, expireAt, accessToken);
       localStorage.setItem('rt', refreshToken);
-    } catch (error) {
-      clearUserData();
-      localStorage.clear();
-      alert('로그인 오류가 발생했습니다. 다시 로그인을 진행해주세요.');
-      navigate('/');
+    } catch (error: any) {
+      throw new Error('토큰 재발급 에러 발생');
     }
-    return;
   }
 
   useEffect(() => {
-    const userData = fetchUserData();
-    if (!userData) {
-      clearUserData();
-      localStorage.clear();
-      navigate('/');
-      return;
-    }
-    const { role, expireAt, accessToken } = userData;
+    if (isAuthChecked) return; // 이미 인증을 체크한 경우 중복 실행 방지
 
-    if (!role || !accessToken || !expireAt) { // 역할이 없으면 로그인 페이지로 리다이렉트
+    const { role, expireAt, accessToken } = fetchUserData();
+
+    if (!role || !accessToken || !expireAt) {
       alert("로그인 오류가 발생했습니다. 다시 로그인을 진행해주세요.");
       clearUserData();
       localStorage.clear();
       navigate('/');
+      setIsAuthChecked(true);
       return;
     }
 
@@ -49,23 +42,35 @@ const useCheckAuth = (pageRole: string) => {
       clearUserData();
       localStorage.clear();
       navigate('/');
+      setIsAuthChecked(true);
       return;
     }
 
     const expireDate = new Date(expireAt);
     const nowDate = new Date();
-    if (expireDate < nowDate) { // 토큰이 만료 됐다면
-      refreshAccessToken(refreshTKOG);
+    if (expireAt && expireDate < nowDate) {
+      try {
+        refreshAccessToken(refreshTKOG);
+        setIsAuthChecked(true);
+      } catch (error: any) {
+        clearUserData();
+        localStorage.clear();
+        alert('로그인 오류가 발생했습니다. 다시 로그인을 진행해주세요.');
+        navigate('/');
+      }
+      return;
     }
 
-    if (role !== pageRole) {
+    if (role && role !== pageRole) {
       alert('로그인 오류가 발생했습니다. 다시 로그인을 진행해주세요.');
       clearUserData();
       localStorage.clear();
       navigate('/');
+      setIsAuthChecked(true);
       return;
     }
   }, []);
+
 }
 
 export default useCheckAuth;
